requirements.txt
## The following requirements were added by pip freeze:
blis==0.2.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==7.1.1
cryptography==2.9
cymem==2.0.3
decorator==4.4.2
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.2
Pint==0.11
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
pycparser==2.20
pymongo==3.10.1
python-dateutil==2.7.5
python-telegram-bot==12.5.1
pytz==2019.3
PyYAML==3.13
requests==2.23.0
six==1.14.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==1.0.2
thinc==7.0.8
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.8
wasabi==0.6.0
Werkzeug==1.0.1
